start of constraint system.  READ NOTE:
Note: I don't know if I will ever be able to even finish this.  I have no idea how to network the physics this complicated, let alone even in this engine.  Networking precise physics such as hinges is probably close to impossible for an engine like this.

This code will most likely not compile, as it is not even stable.  This is just a start of a system that works with bullet.